🏗️ Moved up myExp component to test build
